requirements.txt
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
filelock==3.9.0
Flask==2.2.2
huggingface-hub==0.12.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
numpy==1.24.1
packaging==23.0
Pillow==9.4.0
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
timm==0.6.12
tokenizers==0.13.2
torch==1.13.1+cpu
torchaudio==0.13.1+cpu
torchvision==0.14.1+cpu
tqdm==4.64.1
transformers==4.26.0
typing_extensions==4.4.0
urllib3==1.26.14
Werkzeug==2.2.2
